avoid code duplication
