+++ test/python/visualization/visualization.py
@@ -18,6 +18,8 @@ Useful for refactoring purposes."""
 import os
 import unittest
 import matplotlib
+from filecmp import cmp as cmpfile
+from shutil import copyfile
 
 from qiskit.test import QiskitTestCase
 
@@ -42,6 +44,17 @@ class QiskitVisualizationTestCase(QiskitTestCase):
         with open(current, encoding=encoding) as cur, \
                 open(expected, encoding=encoding) as exp:
             self.assertEqual(cur.read(), exp.read())
+
+    def assertEqualToReference(self, result):
+        reference = path_to_diagram_reference(os.path.basename(result))
+        if not os.path.exists(result):
+            raise self.failureException('Result file was not generated.')
+        if not os.path.exists(reference):
+            copyfile(result, reference)
+        if cmpfile(reference, result):
+            os.remove(result)
+        else:
+            raise self.failureException('Result and reference do not match.')
 
     def assertImagesAreEqual(self, current, expected, diff_tolerance=0.001):
         """Checks if both images are similar enough to be considered equal.
